+++ Src/FluentAssertions/Primitives/StringAssertions.cs
@@ -315,7 +315,40 @@ namespace FluentAssertions.Primitives
         {
             Guard.ThrowIfArgumentIsNull(regularExpression, nameof(regularExpression), "Cannot match string against <null>. Provide a regex pattern or use the BeNull method.");
 
-            if (regularExpression.Length == 0)
+            Regex regex;
+            try
+            {
+                regex = new Regex(regularExpression);
+            }
+            catch (ArgumentException)
+            {
+                Execute.Assertion.FailWith("Cannot match {context:string} against {0} because it is not a valid regular expression.",
+                    regularExpression);
+                return new AndConstraint<TAssertions>((TAssertions)this);
+            }
+
+            return MatchRegex(regex, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that a string matches a regular expression.
+        /// </summary>
+        /// <param name="regularExpression">
+        /// The regular expression with which the subject is matched.
+        /// </param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> MatchRegex(Regex regularExpression, string because = "", params object[] becauseArgs)
+        {
+            Guard.ThrowIfArgumentIsNull(regularExpression, nameof(regularExpression), "Cannot match string against <null>. Provide a regex pattern or use the BeNull method.");
+
+            var regexStr = regularExpression.ToString();
+            if (regexStr.Length == 0)
             {
                 throw new ArgumentException("Cannot match string against an empty string. Provide a regex pattern or use the BeEmpty method.", nameof(regularExpression));
             }
@@ -324,21 +357,13 @@ namespace FluentAssertions.Primitives
                 .ForCondition(!(Subject is null))
                 .UsingLineBreaks
                 .BecauseOf(because, becauseArgs)
-                .FailWith("Expected {context:string} to match regex {0}{reason}, but it was <null>.", regularExpression);
+                .FailWith("Expected {context:string} to match regex {0}{reason}, but it was <null>.", regexStr);
 
-            try
-            {
-                Execute.Assertion
-                .ForCondition(Regex.IsMatch(Subject, regularExpression))
+            Execute.Assertion
+                .ForCondition(regularExpression.IsMatch(Subject))
                 .BecauseOf(because, becauseArgs)
                 .UsingLineBreaks
-                .FailWith("Expected {context:string} to match regex {0}{reason}, but {1} does not match.", regularExpression, Subject);
-            }
-            catch (ArgumentException)
-            {
-                Execute.Assertion
-                    .FailWith("Cannot match {context:string} against {0} because it is not a valid regular expression.", regularExpression);
-            }
+                .FailWith("Expected {context:string} to match regex {0}{reason}, but {1} does not match.", regexStr, Subject);
 
             return new AndConstraint<TAssertions>((TAssertions)this);
         }
@@ -360,7 +385,40 @@ namespace FluentAssertions.Primitives
         {
             Guard.ThrowIfArgumentIsNull(regularExpression, nameof(regularExpression), "Cannot match string against <null>. Provide a regex pattern or use the NotBeNull method.");
 
-            if (regularExpression.Length == 0)
+            Regex regex;
+            try
+            {
+                regex = new Regex(regularExpression);
+            }
+            catch (ArgumentException)
+            {
+                Execute.Assertion.FailWith("Cannot match {context:string} against {0} because it is not a valid regular expression.",
+                    regularExpression);
+                return new AndConstraint<TAssertions>((TAssertions)this);
+            }
+
+            return NotMatchRegex(regex, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that a string does not match a regular expression.
+        /// </summary>
+        /// <param name="regularExpression">
+        /// The regular expression with which the subject is matched.
+        /// </param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotMatchRegex(Regex regularExpression, string because = "", params object[] becauseArgs)
+        {
+            Guard.ThrowIfArgumentIsNull(regularExpression, nameof(regularExpression), "Cannot match string against <null>. Provide a regex pattern or use the NotBeNull method.");
+
+            var regexStr = regularExpression.ToString();
+            if (regexStr.Length == 0)
             {
                 throw new ArgumentException("Cannot match string against an empty regex pattern. Provide a regex pattern or use the NotBeEmpty method.", nameof(regularExpression));
             }
@@ -369,21 +427,13 @@ namespace FluentAssertions.Primitives
                 .ForCondition(!(Subject is null))
                 .UsingLineBreaks
                 .BecauseOf(because, becauseArgs)
-                .FailWith("Expected {context:string} to not match regex {0}{reason}, but it was <null>.", regularExpression);
+                .FailWith("Expected {context:string} to not match regex {0}{reason}, but it was <null>.", regexStr);
 
-            try
-            {
-                Execute.Assertion
-                    .ForCondition(!Regex.IsMatch(Subject, regularExpression))
-                    .BecauseOf(because, becauseArgs)
-                    .UsingLineBreaks
-                    .FailWith("Did not expect {context:string} to match regex {0}{reason}, but {1} matches.", regularExpression, Subject);
-            }
-            catch (ArgumentException)
-            {
-                Execute.Assertion.FailWith("Cannot match {context:string} against {0} because it is not a valid regular expression.",
-                    regularExpression);
-            }
+            Execute.Assertion
+                .ForCondition(!regularExpression.IsMatch(Subject))
+                .BecauseOf(because, becauseArgs)
+                .UsingLineBreaks
+                .FailWith("Did not expect {context:string} to match regex {0}{reason}, but {1} matches.", regexStr, Subject);
 
             return new AndConstraint<TAssertions>((TAssertions)this);
         }